requirements.txt
feedparser>=6.0.11
python-dateutil>=2.9.0.post0
rapidfuzz>=3.6.2
requests>=2.32.3
